pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["andreschererkunz <andreschererkunzt@gmail.com>"]
readme = "README.md"

packages = [
  { include = "core" },
  { include = "config" }
]

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.2,<6.0"
djangorestframework = ">=3.16.0,<4.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
pillow = ">=11.1.0,<12.0.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
gunicorn = ">=23.0.0,<24.0.0"
dj-database-url = ">=2.3.0,<3.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"